cloud_function_pubsub/requirements.txt
google-cloud-bigquery==3.2.0

